Moved span.hpp.
