+++ prototype2/common/span.hpp
